refactor(tui): split event_form.go below 1000 lines
Move constructors and save logic into sibling files in
package tui. No behavior change.

Related: #652
